Refactor: changed importation of tools in their tests
